+++ packages/base/src/components/Avatar/Avatar.tsx
@@ -3,12 +3,12 @@ import { StyledAvatar, StyledAvatarBadge, StyledAvatarContainer, StyledAvatarIma
 import type { AvatarProps } from './Avatar.types';
 import AvatarImageNone from '../../assets/mocks/AvatarNone.png';
 
-const Avatar = ({ src, badge, ...rest }: AvatarProps) => {
+const Avatar = ({ src, badge, onError, ...rest }: AvatarProps): JSX.Element => {
   return (
     <StyledAvatarContainer>
       <StyledAvatar data-shoplflow={'Avatar'} {...rest}>
         {/*eslint-disable-next-line @typescript-eslint/no-unsafe-assignment*/}
-        <StyledAvatarImage src={(src ?? AvatarImageNone) as string} />
+        <StyledAvatarImage src={(src ?? AvatarImageNone) as string} onError={onError} />
       </StyledAvatar>
       <StyledAvatarBadge>{badge}</StyledAvatarBadge>
     </StyledAvatarContainer>