+++ packages/base/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @shoplflow/base
 
+## 0.42.9
+
+### Patch Changes
+
+- [#631](https://github.com/shopl/shoplflow/pull/631) [`c49053c13a42eb84db31a401318c639959556593`](https://github.com/shopl/shoplflow/commit/c49053c13a42eb84db31a401318c639959556593) Thanks [@ayaan0209](https://github.com/ayaan0209)! - SearchBar maxLength + onChange 제대로 안되는 현상 수정
+
+## 0.42.8
+
+### Patch Changes
+
+- [#629](https://github.com/shopl/shoplflow/pull/629) [`a6beacc57618fbd92478c4fbf52b4a5e57c3bf99`](https://github.com/shopl/shoplflow/commit/a6beacc57618fbd92478c4fbf52b4a5e57c3bf99) Thanks [@Eunseo-jo](https://github.com/Eunseo-jo)! - update(component): selectInputButton 컴포넌트 Clear 버튼 조건부 렌더링 추가했습니다
+
+## 0.42.7
+
+### Patch Changes
+
+- [#627](https://github.com/shopl/shoplflow/pull/627) [`c05f19391adebdc67a179ba8ffb9fb6f828a5c26`](https://github.com/shopl/shoplflow/commit/c05f19391adebdc67a179ba8ffb9fb6f828a5c26) Thanks [@velo-kim](https://github.com/velo-kim)! - Update(component): NumberCombobox Dropdown 영역 height 동적 변경
+
 ## 0.42.6
 
 ### Patch Changes